+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed `compas.data.Data.dtype` to `compas.data.Data.__dtype__`.
 * Changed `compas.data.Data.from_data` to `compas.data.Data.__from_data__`.
 * Changed `compas.tolerance.Tolerance` into singleton.
+* Changed `compas_rhino.geometry.curves.nursb.RhinoNurbsCurve` to use private data API.
+* Changed `compas_rhino.geometry.surfaces.nursb.RhinoNurbsSurface` to use private data API.
 
 ### Removed